Use of hooks in the component and responsive in the canvas
